+++ src/EDD_Translations_Updater/Base.php
@@ -1,14 +1,14 @@
 <?php
 /**
- * EDD Software Licensing Translations Updater
+ * EDD Translations Updater
  *
- * @package   Fragen\EDD_SL_Translations_Updater
+ * @package   Fragen\EDD_Translations_Updater
  * @author    Andy Fragen
  * @license   MIT
- * @link      https://github.com/afragen/edd-sl-translations-updater
+ * @link      https://github.com/afragen/edd-translations-updater
  */
 
-namespace Fragen\EDD_SL_Translations_Updater;
+namespace Fragen\EDD_Translations_Updater;
 
 /*
  * Exit if called directly.
@@ -22,7 +22,7 @@ if ( ! defined( 'WPINC' ) ) {
  *
  * Update WordPress language packs from a git hosted repo.
  *
- * @package Fragen\EDD_SL_Translations_Updater
+ * @package Fragen\EDD_Translations_Updater
  * @author  Andy Fragen
  */
 trait Base {
@@ -79,11 +79,13 @@ trait Base {
 		$repos = $this->config;
 		$keys  = array_keys( $repo );
 		array_pop( $keys ); // remove type.
-		$slug = array_filter( $keys, function( $e ) use ( $repos ) {
-			if ( ! isset( $repos[ $e ] ) ) {
-				return $e;
+		$slug = array_filter(
+			$keys, function( $e ) use ( $repos ) {
+				if ( ! isset( $repos[ $e ] ) ) {
+					return $e;
+				}
 			}
-		} );
+		);
 
 		return array_pop( $slug );
 	}