+++ CHANGES.md
@@ -1,5 +1,10 @@
 #### [unreleased]
 
+#### 1.2.0 / 2018-06-28
+* update `class Autoload`
+* update for GitLab API v4
+* renamed to **EDD Translations Updater**
+
 #### 1.1.0 / 2018-01-07
 * No need to use `trait API` in `trait Base` and vica versa